+++ src/states/pause.rb
@@ -3,12 +3,28 @@ require 'state_manager'
 class Pause < StateManager::State
   def initialize(superstate)
     @superstate = superstate
+
+    @current_option = 0
+
+    @arrow_y = 250
   end
 
   def update(_dt)
+    if key_pressed?(KEY_UP) && @current_option > 0
+      @current_option -= 1
+      @arrow_y = 250
+    elsif key_pressed?(KEY_DOWN) && @current_option < (1)
+      @current_option += 1
+      @arrow_y = 300
+    end
+
     return nil unless key_pressed?(KEY_ENTER)
 
-    @superstate.substate = :playing
+    if @current_option.zero?
+      @superstate.substate = :playing
+    else
+      StateManager.set_state(Menu.new)
+    end
   end
 
   def draw
@@ -18,5 +34,13 @@ class Pause < StateManager::State
              .draw(colour: Colour.new(0, 0, 0, 100))
 
     draw_text('-- PAUSE --', 255, 150, 48, WHITE)
+    Rectangle.new(295, @arrow_y, 10, 30).draw(colour: WHITE)
+
+    draw_buttons
+  end
+
+  def draw_buttons
+    draw_text('Resume', 320, 250, 38, WHITE)
+    draw_text('Menu', 320, 300, 38, WHITE)
   end
 end